+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v ?.?.? - ?? ??? ????
 - Removed [[nodiscard]] from ModelLoader::draw() to silence warnings.  Currently method always returns true so the return value isn't applicable.
 - Changed format of extensions listing header
+- Added abstract PerspectiveCamera class to house aspect ratio, field of view, near/far clip planes.  When any is updated, the perspective projection matrix is recomputed
 
 ## v 5.2.0 - 15 Aug 2024
 - OpenGLEngine tracks the state of the Caps Lock and Num Lock keys during key callbacks